% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{est_copula}
\alias{est_copula}
\title{Estimation of the Gaussian copula model with the exchangable correlation matrix}
\usage{
est_copula(m_marginal_model, data, id_name = "id", N_name = "N")
}
\description{
Estimation of the Gaussian copula model with the exchangable correlation matrix
}
