% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{est_y_marginal}
\alias{est_y_marginal}
\title{Estimation of the marginal outcome model}
\usage{
est_y_marginal(formula.Y, data, A_level = 1, A_name = "A")
}
\description{
Estimation of the marginal outcome model
}
