% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mediate_nonpar_est1}
\alias{mediate_nonpar_est1}
\title{Mediation Analysis based on the machine learning estimator under original EIFs}
\usage{
mediate_nonpar_est1(data, formula.M, formula.Y, is_stab, is_true_copula = 1)
}
\arguments{
\item{data}{the dataset}

\item{formula.M}{formula for the mediator model}

\item{formula.Y}{formula for the outcome model}

\item{is_stab}{stabilization or not (1 yes, 0 no)}
}
\value{
Point and interval estimates of NIE, NDE, and SME
}
\description{
Mediation Analysis based on the machine learning estimator under original EIFs
}
