% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mediate_par_est1}
\alias{mediate_par_est1}
\title{Point estimation of mediation effects based on the semiparametric doubly robust estimator under original EIFs}
\usage{
mediate_par_est1(data, formula.M, formula.Y, is_stab, is_true_copula = 1)
}
\arguments{
\item{data}{the dataset}

\item{formula.M}{formula for the mediator model}

\item{formula.Y}{formula for the outcome model}

\item{is_stab}{stabilization or not (1 yes, 0 no)}
}
\value{
Point Estimates of NIE, NDE, and SME
}
\description{
Point estimation of mediation effects based on the semiparametric doubly robust estimator under original EIFs
}
