% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mediate_par_est1_boot}
\alias{mediate_par_est1_boot}
\title{Mediation Analysis based on the semiparametric doubly robust estimator under original EIFs}
\usage{
mediate_par_est1_boot(
  data,
  formula.M,
  formula.Y,
  is_stab,
  B = 100,
  is_true_copula = 1
)
}
\arguments{
\item{data}{the dataset}

\item{formula.M}{formula for the mediator model}

\item{formula.Y}{formula for the outcome model}

\item{is_stab}{stabilization or not (1 yes, 0 no)}

\item{B}{number of bootstrap samples}
}
\value{
Point and interval estimates of NIE, NDE, and SME
}
\description{
Mediation Analysis based on the semiparametric doubly robust estimator under original EIFs
}
