% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mediate_par_est2}
\alias{mediate_par_est2}
\title{Point estimation of mediation effects based on the semiparametric doubly robust estimator under reparameterized EIFs}
\usage{
mediate_par_est2(
  data,
  formula.A,
  formula.M,
  formula.Ma,
  formula.Y,
  formula.Ya,
  formula.Yb,
  is_stab
)
}
\arguments{
\item{data}{the dataset}

\item{formula.A}{formula for the treatment model s(a,m,c,n)}

\item{formula.M}{formula for the mediator model kappa_j(a,mj,c,n)}

\item{formula.Ma}{formula for the mediator model k_j^star(a,m,c,n)}

\item{formula.Y}{formula for the outcome model eta_j(a,m,c,n)}

\item{formula.Ya}{formula for the outcome model eta_j^star(a,a*,c,n)}

\item{formula.Yb}{formula for the outcome model eta_j^dagger(a,a*,mj,c,n)}

\item{is_stab}{stabilization or not (1 yes, 0 no)}
}
\value{
Point Estimates of NIE, NDE, and SME
}
\description{
Point estimation of mediation effects based on the semiparametric doubly robust estimator under reparameterized EIFs
}
