% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mediateCRT_est1}
\alias{mediateCRT_est1}
\title{Mediation analysis based on the machine learning estimator under reparameterized EIFs}
\usage{
mediateCRT_est1(
  data,
  id_name,
  N_name,
  A_name,
  M_name,
  Y_name,
  X_names,
  V_names,
  is_stab,
  B
)
}
\arguments{
\item{data}{the dataset}

\item{id_name}{cluster id}

\item{N_name}{cluster size}

\item{A_name}{treatment}

\item{M_name}{mediator}

\item{Y_name}{outcome}

\item{X_names}{individual-level covariates}

\item{V_names}{cluster-level covariates (if no cluster covariates, then set V_names=NULL)}

\item{is_stab}{stabilization or not (1 yes, 0 no)}

\item{B}{number of bootstrap}
}
\value{
Point and interval estimates of NIE, NDE, and SME
}
\description{
Mediation analysis based on the machine learning estimator under reparameterized EIFs
}
