% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MF_tau_par_est1}
\alias{MF_tau_par_est1}
\title{Estimation of the mediation functional tau based on the semiparametric doubly robust estimator under original EIFs}
\usage{
MF_tau_par_est1(
  data,
  formula.M,
  formula.Y,
  m_res,
  y1_res,
  y0_res,
  copula_res,
  is_stab = 1
)
}
\arguments{
\item{data}{the dataset}

\item{formula.M}{formula for the mediator model}

\item{formula.Y}{formula for the outcome model}

\item{m_res}{regression output from the marginal mediator model}

\item{y1_res}{regression output from the marginal outcome model under the treated clusters}

\item{y0_res}{regression output from the marginal outcome model under the control clusters}

\item{copula_res}{output from the copula model}

\item{is_stab}{stabilization or not (1 yes, 0 no)}
}
\value{
The estimates of tau
}
\description{
Estimation of the mediation functional tau based on the semiparametric doubly robust estimator under original EIFs
}
