% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MF_theta_nonpar_est1}
\alias{MF_theta_nonpar_est1}
\title{Estimation of the mediation functional theta(a,a*) based on the machine learning estimator under original EIFs}
\usage{
MF_theta_nonpar_est1(
  data,
  a1 = 1,
  a0 = 0,
  formula.M,
  formula.Y,
  is_stab = 1,
  V = 5,
  is_true_copula = 1,
  myseed = 2023
)
}
\arguments{
\item{data}{the dataset}

\item{a1}{the value *a* in *\theta(a,a^\star)*}

\item{a0}{the value *a^\star* in *\theta(a,a^\star)*}

\item{formula.M}{formula for the mediator model}

\item{formula.Y}{formula for the outcome model}

\item{is_stab}{stabilization or not (1 yes, 0 no)}

\item{V}{number of groups based on the cross-fitting procedure}
}
\value{
The estimates of theta(a,a*)
}
\description{
Estimation of the mediation functional theta(a,a*) based on the machine learning estimator under original EIFs
}
