% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MF_theta_par_est2}
\alias{MF_theta_par_est2}
\title{Estimation of the mediation functional theta(a,a*) based on the semiparametric doubly robust estimator under reparameterized EIFs}
\usage{
MF_theta_par_est2(
  data,
  a1 = 1,
  a0 = 0,
  formula.A,
  formula.Y,
  formula.Ya,
  is_stab = 1
)
}
\arguments{
\item{data}{the dataset}

\item{a1}{the value *a* in *\theta(a,a^\star)*}

\item{a0}{the value *a^\star* in *\theta(a,a^\star)*}

\item{formula.A}{formula for the conditional treatment model s(a,m,c,n)}

\item{formula.Y}{formula for the outcome model eta_j(a,m,c,n)}

\item{formula.Ya}{formula for the outcome model eta_j^star(a,a*,c,n)}

\item{is_stab}{stabilization or not (1 yes, 0 no)}
}
\value{
The estimates of theta(a,a*)
}
\description{
Estimation of the mediation functional theta(a,a*) based on the semiparametric doubly robust estimator under reparameterized EIFs
}
