% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Example dataset: sim_data}
\format{
A data frame with 100 clusters and 5 variables:
\describe{
  \item{id}{cluster ID}
  \item{N}{cluster size}
  \item{A}{treatment indicator}
  \item{C}{a cluster-level covariate}
  \item{X}{a individual-level covariate}
  \item{M}{mediator}
  \item{Y}{outcome}
}
}
\source{
Simulated data for demonstration
}
\usage{
sim_data
}
\description{
A brief description of what this dataset contains and its context.
}
\keyword{datasets}
